pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "finagent-mcp-server"
version = "0.1.0"
description = "FinAgent Model Context Protocol Server for AI Agents with Crypto Payment Support"
authors = [
    {name = "FinAgent", email = "admin@finagent.com"},
]
dependencies = [
    # Official MCP Python SDK
    "mcp[cli]>=1.11.0",
    
    # Web and API dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    
    # Data validation and settings
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    
    # Crypto and blockchain
    "cryptography>=41.0.0",
    "web3>=6.11.0",
    "tronpy>=0.4.0",
    
    # Utilities
    "loguru>=0.7.0",
    "psutil>=5.9.0",
    
    # Authentication
    "PyJWT>=2.8.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.26.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 